+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -18,6 +18,7 @@
 #include "Rule10p8Check.h"
 #include "Rule12p3Check.h"
 #include "Rule14p4Check.h"
+#include "Rule15p1Check.h"
 #include "Rule15p5Check.h"
 #include "Rule15p6Check.h"
 #include "Rule15p7Check.h"
@@ -62,6 +63,8 @@ public:
         "misrac-rule-12p3");
     CheckFactories.registerCheck<Rule14p4Check>(
         "misrac-rule-14p4");
+    CheckFactories.registerCheck<Rule15p1Check>(
+        "misrac-rule-15p1");
     CheckFactories.registerCheck<Rule15p5Check>(
         "misrac-rule-15p5");
     CheckFactories.registerCheck<Rule15p6Check>(